Improved multi-platform compatibility. 

The default MinGW compiler from http://www.mingw.org/ does not support threads. 
Windows developers should replace their MinGW installation with one like this: http://sourceforge.net/projects/mingwbuilds/files/host-windows/releases/4.8.1/64-bit/threads-posix/sjlj/x64-4.8.1-release-posix-sjlj-rev5.7z/download

+++ gamed/include/Map.h
@@ -21,7 +21,7 @@ public:
    Map(Game* game) : game(game) { }
    
    virtual ~Map() { }
-   virtual void update(unsigned int diff);
+   virtual void update(long long diff);
    Object* getObjectById(uint32 id);
    void addObject(Object* o);
